Refactor: clean up and track deleted files
