+++ src/main/kotlin/org/conservationco/asana/exception/CustomFieldException.kt
@@ -1,3 +1,3 @@
-package org.conservationco.asana.exceptions
+package org.conservationco.asana.exception
 
 class CustomFieldException(message: String) : IllegalArgumentException(message)
