--- src/main/kotlin/org/conservationco/asana/serialization/customfield/CustomFieldException.kt
@@ -1,3 +1,0 @@
-package org.conservationco.asana.serialization.customfield
-
-class CustomFieldException(message: String) : IllegalArgumentException(message)
